project/requirements.txt
aiofiles==0.6.0
celery==5.2.0
fastapi==0.70.0
flower
pytest==6.2.4
redis==4.1.4
requests==2.25.1
uvicorn==0.15.0
SQLAlchemy==1.4.26
pymysql==1.0.2
opencv-python
ffmpeg-python